requirements.txt
requests==2.32.5
beautifulsoup4==4.14.3
httpx==0.28.1
aiohttp==3.13.2
selectolax==0.4.6
pandas==2.3.3
pyarrow==22.0.0
google-cloud-bigquery==3.38.0
google-cloud-storage==3.7.0
tqdm==4.67.1
python-dotenv==1.2.1
python-dateutil==2.9.0
selenium==4.39.0
spacy==3.8.11
langdetect==1.0.9
wheel==0.45.1
langchain==1.2.0
langgraph==1.0.5
mcp==1.24.0
google-cloud-aiplatform==1.132.0
sentence-transformers==5.2.0
fastapi==0.128.0
uvicorn==0.40.0
slowapi==0.1.9
python-multipart==0.0.21
lightgbm==4.6.0
scikit-learn==1.8.0
xgboost==3.1.2
umap-learn==0.5.9.post2
functions-framework==3.10.0
matplotlib==3.10.8
seaborn==0.13.2
db-dtypes==1.5.0
pytest==9.0.2
ollama==0.6.1
chromadb==1.5.5
mysql-connector-python==9.6.0
langchain-experimental==0.4.1
langchain-ollama==1.0.1
opentelemetry-api==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-instrumentation-fastapi==0.60b1
opentelemetry-exporter-gcp-trace==1.11.0
opentelemetry-exporter-gcp-monitoring==1.11.0a0
prometheus-client==0.23.1